ForrestHub-app/requirements.txt
flask~=3.1.0
flask-socketio~=5.5.1
gevent-websocket~=0.10.1
requests~=2.32.3
websocket-client~=1.8.0
flask-cors~=5.0.1
pyinstaller~=6.12.0
pyopenssl~=25.0.0
pyinstaller~=6.12.0
nanoid~=2.0.0
eventlet~=0.39.1
python-dotenv~=1.0.1
click~=8.1.8
